Release 5.0.0.
Unlike 4.10.0 that was just released this is Python 3 only (drop 3.4
support and bump the compatibility with IPython to IPython 5+ required).

It also have all the machinery to have async/await top level.

I believe the modification I've done to the recipe are sufficient, but
I'd appreciate if one can be extra careful